// React/MajorProject/crypto-pilot/src/features/market/types/coin.ts
export interface Coin {
  id: string;
  symbol: string;
  name: string;
  image: string;

  current_price: number;

  market_cap: number;

  market_cap_rank: number;

  total_volume: number;

  price_change_percentage_24h: number;

  high_24h: number;

  low_24h: number;

  circulating_supply: number;
}